Support merging of values from default profile in config module.
The Config() object takes a new constructor parameter merge_default.
When this paramter is True, config values of types list or dict will be
returned as the combination of values from the currently selected
profile as well as the default profile.  This allows a profile to
"inherit" settings from the default profile.