Simplify feed configuration and support additional feed options.
Previously, feeds could be defined as a list of locators or a dict of
locators mapped to names, or a list of dicts.  This was all getting
unwieldy, so it's been simplified to be a dict of entries matfching one
of the following forms:

    "locator":

    "locator": "Name of feed"

    "locator": name: "Name of feed"
        option1: "foo"
        optino2: "bar"

This means feeds are now uniquely identified by their locator instead of
their name, which makes more sense.  This syntax also allows setting
options for each feed in the configuration, which will be used by the
translation feature to suggest a source language for a feed's content.