Eliminate dinstinction between Helper and Downloader objects.
At one time, it seemed to make sense to have seperate Helper programs
and Downloader programs, but really the only distinction that matters
here is "does this app play media" or "does it just fetch the media".
The helpers we care about (youtube-dl and streamink) can both download,
so we'll just call them downloaders along with wget, curl, etc.